MapElites refactor.
